requirements.txt
numpy>=1.26.4
pandas>=2.2.0
scipy>=1.13.0
yfinance>=0.2.40
alpaca-trade-api>=3.1.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0